Share safe_path between the two Sonar scripts
The quality gate moved from a coverage failure to 4.3% duplication on new
code, against a 3% limit. Cause: safe_path() was added to both
fetch_sonar_issues.py and sonar_to_sarif.py as identical copies. I judged
that duplication too small to matter when adding it; at ~15 lines across
two files it was not.

Extracted to scripts/sonar_paths.py. Both scripts run as
`python3 scripts/<name>.py`, which puts scripts/ on sys.path, so a plain
import resolves without any path juggling -- the same reason the test file
already does sys.path.insert.

Verified: 26/26 tests pass, both scripts still run standalone from the
repository root, and the generated SARIF is unchanged and still validates
against the OASIS 2.1.0 schema.
