+++ scripts/sonar_to_sarif.py
@@ -25,6 +25,8 @@ import sys
 from pathlib import Path
 from typing import Any, Iterable
 
+from sonar_paths import safe_path
+
 SARIF_VERSION = "2.1.0"
 SARIF_SCHEMA = "https://json.schemastore.org/sarif-2.1.0.json"
 
@@ -188,23 +190,6 @@ def load_issues(raw: str) -> list[dict[str, Any]]:
     if isinstance(data, dict):
         return list(data.get("issues") or [])
     return list(data)
-
-
-def safe_path(value: str, *, must_exist: bool = False) -> Path:
-    """Resolve a caller-supplied path, refusing anything outside the working tree.
-
-    These scripts only ever read and write artefacts inside the checkout, so a path that escapes
-    it (`../../etc/passwd`) is always a mistake or an attack rather than a legitimate use. Anchor
-    to the working directory and reject the rest, instead of trusting whatever the CLI was given.
-    """
-    base = Path.cwd().resolve()
-    candidate = Path(value)
-    candidate = candidate.resolve() if candidate.is_absolute() else (base / candidate).resolve()
-    if not candidate.is_relative_to(base):
-        raise SystemExit(f"error: refusing a path outside {base}: {value}")
-    if must_exist and not candidate.is_file():
-        raise SystemExit(f"error: no such file: {value}")
-    return candidate
 
 
 def main(argv: list[str] | None = None) -> int:
